% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{coclust_mat}
\alias{coclust_mat}
\title{Compute the posterior similarity matrix (probability that i and j are in same cluster).}
\usage{
coclust_mat(n, z)
}
\arguments{
\item{n}{total number of subjejcts}

\item{z}{a matrix of posterior samples, rows for subejcts, columns for MCMC samples kept across
multiple iterations}
}
\value{
a matrix of co-clustering frequencies
}
\description{
This function is to evaluate the recovery of clusters
}
