% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{create.data}
\alias{create.data}
\title{Create Time Delay Embedding}
\usage{
create.data(ts, embed = 10)
}
\arguments{
\item{ts}{the data structure with the time series data}

\item{embed}{the size of the embed, defaults to 10}
}
\value{
a data frame with the time delay embedding of the time series
}
\description{
Time series forecasting models usually assume the existence of a degree of correlation between successive values of the series. A form of modelling such correlation is to use previous values of the series as predictors of future value(s), in a procedure known as time delay embedding. This process allows the us of standard regression tools on time series forecasting tasks.
}
\examples{

}
