% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cuthill_mckee}
\alias{cuthill_mckee}
\title{Cuthill McKee (CM) algorithm}
\usage{
cuthill_mckee(x)
}
\arguments{
\item{x}{Input matrix}
}
\value{
Band matrix
}
\description{
Transform sparse matrix into a band matrix
}
\examples{

L0 <- 100   # dimension of measurements.
M0 <- 30    # true dimension of latent states.
frac <- 0.05
NITER <- 1
maxlen_simu <- rep(NA,NITER)
for (iter in 1:NITER){
  Q    <- simulate_Q(M0,L0,p = frac)
  simu <- list(Q=Q)
  block_ind_list <- mblock(simu$Q,PLOT = TRUE)
  maxlen_simu[iter] <- max(unlist(lapply(block_ind_list,length)))
}
#hist(maxlen_simu)
table(maxlen_simu)
}
\seealso{
\code{\link{mblock}}
}
\author{
Fabian Schmich
}
