% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{distAB}
\alias{distAB}
\title{Frobenius norm: ||AA'-BB'||_F}
\usage{
distAB(A, B)
}
\arguments{
\item{A, B}{two matrixes of the same number of rows}
}
\value{
a positive number
}
\description{
Frobenius norm: ||AA'-BB'||_F
}
