% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{do_update_Q}
\alias{do_update_Q}
\title{determine to update the current element of Q_ml or not}
\usage{
do_update_Q(Q)
}
\arguments{
\item{Q}{a matrix with rows being machines and columns being protein landmarks (dimension)}
}
\value{
a matrix filled with logical values of dimensions identical to Q. TRUE for updating
in constrained Gibbs sampler, FALSE for skipping the updating.
}
\description{
Function to test whether we need to update the current element of Q_ml. This
is needed in the constrained Gibbs sampler.
}
