% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{do_update_Q_one}
\alias{do_update_Q_one}
\title{determine to update the current element of Q_ml or not}
\usage{
do_update_Q_one(Q, k, l)
}
\arguments{
\item{Q}{a matrix with rows being machines and columns being protein landmarks (dimension)}

\item{k, l}{row and column indicators for checking wether to update Q_kl}
}
\value{
logical value. TRUE for updating in constrained Gibbs sampler, FALSE
for skipping the updating.
}
\description{
Function to test whether we need to update the current element of Q_ml. This
is needed in the constrained Gibbs sampler.
}
