% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{equal_unit_vec}
\alias{equal_unit_vec}
\title{check whether a vector is equal to a unit vector with the one at a particular
position}
\usage{
equal_unit_vec(v, k)
}
\arguments{
\item{v}{the vector (a binary vector)}

\item{k}{the index that is being checked if \code{v[k]} is the only one in
vector \code{v}. \code{k} must be smaller than or equal to the length of k}
}
\value{
TRUE if \eqn{v = \mathbf{e}_k}; FALSE otherwise.
}
\description{
This function is used in updating Q matrix if we constrain the updates within
the identifiability assumption
}
\examples{
 equal_unit_vec(c(1,0,0,0,0,0),1)
 equal_unit_vec(c(1,0,0,0,0,0),2)
 equal_unit_vec(c(0,0,2,0,0,0),3)
 equal_unit_vec(c(0,0,1,0,0,0),3)
}
