% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_freq}
\alias{get_freq}
\title{Get frequency count for each integer}
\usage{
get_freq(v, maxk = 20)
}
\arguments{
\item{v}{the vector of integers to be tabulated}

\item{maxk}{maximum integer to search within v}
}
\value{
a vector of probabilities that sum to one
}
\description{
tabulate v by 1:maxk
}
