% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{log_f_logprima}
\alias{log_f_logprima}
\title{compute the deriviative of the log of density of inactive probability given
the previous one (Teh et al., 07')}
\usage{
log_f_logprima(log_p0, alpha, t)
}
\arguments{
\item{log_p0}{log probability (the argument of this function)}

\item{alpha}{hyperparameter for Indian Buffet Process (Infinite version)}

\item{t}{the number of rows in the IBP}
}
\value{
a value corresponding to the derivative of the log density d/du f(log_p0)
}
\description{
This function is used in sampling infinite number of columns in H
}
\seealso{
\code{\link{log_f_logden}} provides an example
}
