% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{log_full0}
\alias{log_full0}
\title{Function to compute the full cluster-specific likelihood given latent variables}
\usage{
log_full0(Y, eta_star, Q, p, theta, psi)
}
\arguments{
\item{Y}{the data for the current cluster (a subset of observations.)}

\item{eta_star}{A matrix of M columns (of machines). Multivariate binary indicators of presence or absence of
protein landmarks (could be a matrix with rows for multiple )}

\item{Q}{Q-matrix}

\item{p}{prevalence parameter for each machine; should be a vector of dimension M.}

\item{theta}{true positive rates}

\item{psi}{true positive rates}
}
\value{
log of marginal likelihood given other model parameters.
}
\description{
This function computes the likelihood WITHOUT integrating over
the distribution of component specific parameter (e.g., machine usage profiles).
This function conditions upon a few model parameters: the true and false positive
rates (theta and psi), the Q matrix and {p}-the prevalence parameter for each machines.
}
