% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{log_pr_Qml_cond_enum}
\alias{log_pr_Qml_cond_enum}
\title{compute the full conditional probability of column Q_l given others}
\usage{
log_pr_Qml_cond_enum(Yl, eta_star, Ql_enumerate, thetal, psil)
}
\arguments{
\item{eta_star}{a matrix of machine usage indicators, rows for clusters, columns for M machines}

\item{Ql_enumerate}{l-th column of the Q matrix}

\item{thetal, psil}{True and false positive rates. Both are vectors of length L}

\item{Y}{data}

\item{p}{prevalence of machines; a vector of length identical to the columns of H.}
}
\value{
a vector of log conditional probability of column Q_l given other unknown parameters,
the dimension will be identical to \code{2^(length(Ql))}.s
}
\description{
Function to compute this log full conditional density, which will be used
in \link{update_Q}
}
