% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mblock}
\alias{mblock}
\title{obtain the indices of rows of Q that form orthogonal blocks}
\usage{
mblock(Q, PLOT = FALSE)
}
\arguments{
\item{Q}{binary Q matrix (M by L), the rows of which are to be split}

\item{PLOT}{plot blocks in the band matrix or not}
}
\value{
a list of indices that form blocks
}
\description{
obtain the indices of rows of Q that form orthogonal blocks
}
\seealso{
\code{\link{cuthill_mckee}}
}
