% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{merge_H_col}
\alias{merge_H_col}
\title{Merge redundant H_star by column}
\usage{
merge_H_col(H_star_redun, Q, VERBOSE = FALSE)
}
\arguments{
\item{H_star_redun}{A binary matrix of rows \code{t_max+3} by \code{m_max};
It may be redundant because it has rows corresponding to empty pseudo-clusters,
and if the entire matrix is non-zero, the zero columns mean inactive latent states}

\item{Q}{Q-matrix - needs to be merged according to H_star_redun.}

\item{VERBOSE}{Default to \code{FALSE}: no print of the merged matrices; Otherwise
set to \code{TRUE}.}
}
\value{
a list of one element named \code{H_star_merge}
}
\description{
This function removes zero columns if thre are more than two rows in H_star
and if \code{H_star_redun} is not trivially all zero
}
