% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{merge_map}
\alias{merge_map}
\title{Merge mapping}
\usage{
merge_map(pseudo_pat, uniq_pat)
}
\arguments{
\item{pseudo_pat}{a vector of character strings for
binary patterns that might not be unique}

\item{uniq_pat}{a vector of character strings of distinct binary codes}
}
\value{
a list \itemize{
\item \code{map} a vector of integer of identical length to \code{pseudo_pat}; takes values
from 1 to \code{length(uniq_pat)}
\item \code{uniq_pat} a matrix of unique binary patterns (# rows =  \code{length(uniq_pat)},
# columns = number of 1/0s for each element in \code{uniq_pat})
}
}
\description{
This function combines pseudo-clusters by the current H_star samples.
}
\examples{

tmp <-  simu$H_star[c(1,1,2,2,2,3,4,5,6,7,8),]
uid <- unique(apply(tmp,1,paste,collapse=""))
merge_map(apply(tmp,1,paste,collapse=""),uid)
}
