% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{metrop_flip}
\alias{metrop_flip}
\title{Metroplis flipping to improve mixing}
\usage{
metrop_flip(x, curr_prob)
}
\arguments{
\item{x}{current value, 0 or 1}

\item{curr_prob}{current probability of being 1}
}
\value{
a binary value
}
\description{
Metroplis flipping to improve mixing
}
\examples{
metrop_flip(1,0.2)


p = 0.3
x <- rep(NA,1000)
x[1] <- 0
for (iter in 2:1000){
  x[iter] <- metrop_flip(x[iter-1],p)
}

plot(x,type="l",main=mean(x))
}
