% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute.R
\name{mfm_coefficients}
\alias{mfm_coefficients}
\title{computing coefficients for mixture of finite mixture model}
\usage{
mfm_coefficients(log_pk, gamma, n, upto)
}
\arguments{
\item{log_pk}{log of prior probability: pr(K=k)}

\item{gamma}{the hyperparameter for symmetric Dirichlet prior for pi_K}

\item{n}{number of observations}

\item{upto}{the upper limit (t<= upto) to compute the coefficient}
}
\value{
a vector of log coefficients in mixture of finite mixtures
}
\description{
Function to compute log(V_N(T)) under given MFM parameters for t = 1:upto.
Pre-computed prior to posterior sampling.
}
