% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{ordered_next}
\alias{ordered_next}
\title{Find the next available cluster ID to use:}
\usage{
ordered_next(mylist)
}
\arguments{
\item{mylist}{a vector of integer cluster IDs}
}
\value{
an index to use next for an observation during complete Gibbs scan
}
\description{
Find the next available cluster ID to use:
}
\examples{
ordered_next(c(1:3,5:10))
ordered_next(c(1:3,5:6,9:10))
}
