% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{ordered_remove}
\alias{ordered_remove}
\title{remove an index from a vector of cluster ID}
\usage{
ordered_remove(index, mylist, t)
}
\arguments{
\item{index}{an integer of cluster ID to be inserted}

\item{mylist}{a vector of integer cluster IDs}

\item{t}{the effective number of clusters}
}
\description{
This function is called when removing a cluster.
}
\examples{
ordered_remove(3,c(1:3,5:10),7)
ordered_remove(3,c(1:10),7)


# some quick examples:

list0 <- c(1:7,0,0,0)
k <- 1
while (list0[k]==k){ k=k+ 1}; cm = k
while (list0[k]==k+1){ k =k+ 1}; ci = k+1
while (list0[k]==k+2) {k =k+ 1};  cj = k+2

a1 <- ordered_remove(3,list0,7)
a2 <- ordered_remove(5,a1,6)
a3 <- ordered_insert(cm,a2,5)

# end of examples.

}
