% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{plot_rewind}
\alias{plot_rewind}
\title{Plot data, design matrix or co-clustering probabilities}
\usage{
plot_rewind(x, simu, type = "data", title_val = type)
}
\arguments{
\item{x}{data, design matrix or coclustering probabilities}

\item{simu}{simulated objects; see \code{\link{simulate_data}}}

\item{type}{default to "data"; could be "design" or "cocluster"}

\item{title_val}{the title of the plot; default is just \code{type}}
}
\value{
a plot
}
\description{
Plot data, design matrix or co-clustering probabilities
}
\examples{
# simulate data:
L0 <- 100   # dimension of measurements.
M0 <- 3     # true dimension of latent states.
K0 <- 8     # true number of clusters.
options_sim0  <- list(N = 100,    # sample size.
                      M = M0,     # true number of machines.
                      L = L0,     # number of antibody landmarks.
                      K = K0,     # number of true components.
                      theta = rep(0.8,L0),  # true positive rates.
                      psi   = rep(0.15,L0), # false positive rates.
                      #alpha1 = 1, # half of the people have the first machine.
                      frac = 0.2, # fraction of positive dimensions (L-2M) in Q.
                      #pop_frac = rep(1/K0,K0) # population prevalences.
                      #pop_frac = (1:K0)/sum(1:K0) # population prevalences.
                      pop_frac = c(rep(2,4),rep(1,4)) # population prevalences.
                      #pop_frac = c(rep(0.75/4,4),rep(0.25/4,4))
)

simu     <- simulate_data(options_sim0, SETSEED=TRUE)
plot_rewind(simu$datmat,simu,"data","Simulated Data")
}
