% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{postprocess_H_Q}
\alias{postprocess_H_Q}
\title{Merge H and Q}
\usage{
postprocess_H_Q(outres, Q = NULL)
}
\arguments{
\item{outres}{output from \code{\link{sampler}}}

\item{Q}{Q can be specified to be known. Default is \code{NULL}.}
}
\value{
a new output with extra elements:
\code{H_star_merge_samp}, \code{Q_merge_samp},
\code{z_sci_samp}, \code{col_merged_H_star_samp}.
}
\description{
NB: need to add another option to skip merge Q when Q is known, in
which case, no col merging, no Q merging, just simple row-wise merging.
Currently function merge_H_Q can deal with this. So just need to use that 
functionality by specifying skip_Q = TRUE.
}
