% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randOverT.R
\name{randOverT}
\alias{randOverT}
\title{Temporal Biased Random Oversampling}
\usage{
randOverT(
  form,
  dat,
  rel = "auto",
  thr.rel = 0.5,
  C.perc = "balance",
  repl = TRUE
)
}
\arguments{
\item{form}{a model formula}

\item{dat}{the original training set (with the unbalanced distribution)}

\item{rel}{is the relevance determined automatically (default: "auto") or provided by the user through a matrix. See examples.}

\item{thr.rel}{is the relevance threshold above which a case is considered as an extreme value}

\item{C.perc}{is a list containing the over-sampling percentage/s to apply to
all/each "class" obtained with the relevance threshold. The percentage represents the percentage of replicas that are added. Replicas of the examples are added randomly in each "class". Moreover, different percentages may be provided for each "class". Alternatively, it may be "balance" (the default) or "extreme", cases where the over-sampling percentages are automatically estimated.}

\item{repl}{is it allowed to perform sampling with replacement (bootstrapping). Defaults to TRUE because if the over-sampling percentage is >2 this is necessary.}
}
\value{
a new training data set resulting from the application of the resampling strategy
}
\description{
Temporal Biased Random Oversampling
}
\examples{
library(rewind)
data(temp)
ds <- create.data(temp,10)
C.perc <- list(4)
overT <- randOverT(V10 ~ ., ds, C.perc=C.perc)
overT.Bal <- randOverT(V10 ~ ., ds, C.perc="balance")
overT.Ext <- randOverT(V10 ~ ., ds, C.perc="extreme")

}
