% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randUnderTPhi.R
\name{randUnderTPhi}
\alias{randUnderTPhi}
\title{Temporal and Relevance Biased Random Undersampling}
\usage{
randUnderTPhi(form, data, rel = "auto", thr.rel = 0.5, C.perc = "balance",
  repl = FALSE)
}
\arguments{
\item{form}{a model formula}

\item{data}{the original training set (with the unbalanced distribution)}

\item{rel}{is the relevance determined automatically (default: "auto") or provided by the user through a matrix. See examples.}

\item{thr.rel}{is the relevance threshold above which a case is considered as an extreme value}

\item{C.perc}{is a list containing the under-sampling percentage/s to apply to all/each "class" obtained with the relevance threshold. This percentage represents the percentage of examples that is maintained in each "class". Examples are randomly removed in each "class". Moreover, different percentages maybe provided for each "class". Alternatively, it may be "balance" or "extreme", cases where the under-sampling percentages are automatically estimated.}

\item{repl}{is it allowed to perform sampling with replacement (bootstrapping)}
}
\value{
a new training data set resulting from the application of the resampling strategy
}
\description{
Temporal and Relevance Biased Random Undersampling
}
\examples{
library(rewind)
data(temp)
ds <- create.data(temp,10)
C.perc <- list(0.5)
underTPhi <- randUnderTPhi(V10 ~ ., ds, C.perc=C.perc)
underTPhi.Bal <- randUnderTPhi(V10 ~ ., ds, C.perc="balance")
underTPhi.Ext <- randUnderTPhi(V10 ~ ., ds, C.perc="extreme")

}
