% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{restricted_gibbs}
\alias{restricted_gibbs}
\title{Restricted Gibbs scan to update assignment indicators in S.}
\usage{
restricted_gibbs(
  Y,
  zsa,
  zsb,
  cia,
  cib,
  cja,
  cjb,
  ni,
  nj,
  i,
  j,
  S,
  ns,
  b,
  active,
  Q,
  p,
  theta,
  psi
)
}
\arguments{
\item{Y}{multivariate binary data (row for subjects, column for dimensions)}

\item{zsa}{cluster assignment indicators prior to updating (transition from in MH)}

\item{zsb}{cluster assignment indicators after updating (transition to in MH; same below.)
both zsa and zsb will be set to zs when active is true, because we want the zs to be
modified/updated.}

\item{cia, cja}{cluster indicator for subject i,j prior to updating}

\item{cib, cjb}{cluster indicator for subject i,j after updating}

\item{ni, nj}{the number of subjects clustered with i (j) - including i (j).}

\item{i, j}{two distinct subjects' indices}

\item{S}{the set of indices clustered with either i or j (including i and j)}

\item{ns}{size of S}

\item{b}{gamma in the Dirichlet distribution in MFM.}

\item{active}{TRUE for performing a split update (both updating cluster indicators and computing the transition probabilities), FALSE otherwise (only for calculating
proposal density; used in q(Z|Z_merge) = q(Z|Z_launch)).}

\item{Q}{a Q matrix of dimension M by L}

\item{p}{a vector of machine (factor) prevalences of length M}

\item{theta}{a vector of true positive rates of length L}

\item{psi}{a vector of false positive rates of length L}
}
\value{
\itemize{
\item \code{log_p} the log probability of transitioning from zsa to zsb (product of all subjects in S);
\item \code{ni,nj} the number of subjects assigned to cluster for i,j
\item \code{zsb} updated assignment indicators for subjects in S
}
}
\description{
Function to perform Gibbs scan only for the observations associated
with either i or j (randomly chosen), but not including the cluster indicator
for i or j. This function does not change the cluster observation i or j currently is in,
because the restricted Gibbs scan only updates other observations in S by randomly
assigning them to the cluster with i or j.
}
