% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewind.R
\docType{package}
\name{rewind}
\alias{rewind}
\title{rewind: \strong{R}constructing \strong{E}tiology \strong{w}ith
B\strong{in}ary \strong{D}ecomposition}
\description{
\code{rewind} is designed for analyzing multivariate binary data via
binary factor analyses, with or without pre-specified number of factors,
and without specifying the number of clusters in the data. It is motivated by analyzing the multivariate presence or absence
of a list of antigens in serum samples collected from autoimmune disease patients.
This package provides a tool for statistical inference of a model that assumes human body generates antibodies to
a small number of protein complexes (or, machines), each comprised of a few important antigens.
}
\details{
\code{rewind} This package implements a Bayesian hierarchical model that represents
observations as aggregation of a few unobserved binary machines where the aggregation
varies by subjects. Our approach is to specify the model likelihood via factorization
into two latent binary matrices: machine profiles and individual factors.
Given latent factorization, we account for inherent errors in measurement using
sensitivities and specificities of protein detection.
We use a prior for the individual factor matrix (Indian Buffet Process for binary matrices) to
encourage a small number of subject clusters each with distinct patterns of active machines.
The posterior distribution for the numbers of patient clusters and machines are
estimated from data and by design tend to concentrate on smaller values.
The posterior distributions of model parameters are estimated via Markov chain
Monte Carlo which makes a list of molecular machine profiles with uncertainty
quantification as well as patient-specific posterior probability of having each machine.
}
\section{Main rewind functions}{

\code{\link{sampler}}
}

\references{
\itemize{
\item This package partly adapts Julia programs used in
Miller, J.W. and Harrison, M.T., 2017. Mixture models with a prior on the number of components.
Journal of the American Statistical Association, pp.1-17. \url{https://github.com/jwmi/BayesianMixtures.jl}
\item Slice sampler for Indian Buffet Process priors: Teh, Y. W., Grur, D., and Ghahramani, Z. (2007).
Stick-breaking construction for the indian buffet process. In Artificial Intelligence and Statistics, pages 556–563.
}
}
\seealso{
\itemize{
\item \url{https://github.com/zhenkewu/rewind} for the source code
and system/software requirements to use \code{rewind} for your data.
}
}
