% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{rewind_dependency}
\alias{rewind_dependency}
\title{The rewind HTML dependency}
\usage{
rewind_dependency()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
\code{\link[=rewind_enable]{rewind_enable()}} adds this automatically. You thus rarely need this
function. Use it when you want the assets before the server function
runs. Use it also when you cannot use \code{insertUI()}.
}
\examples{
rewind_dependency()
}
