% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{rewind_ui}
\alias{rewind_ui}
\title{A scrubbable history rail}
\usage{
rewind_ui(label = "History", max_height = "20rem", class = NULL)
}
\arguments{
\item{label}{The heading above the rail. Use \code{NULL} for no heading.}

\item{max_height}{The CSS height at which the rail starts to scroll.}

\item{class}{More CSS classes for the container element.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}}.
}
\description{
This function draws the history as a vertical list of steps. The newest
step is at the end. The rail shows the current position. A click on a
step moves to that step. The rail gets its data from the server. You thus
do not have to write a \code{render} function.
}
\examples{
rewind_ui()
rewind_ui(label = "Steps", max_height = "12rem")
}
