% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{sampler}
\alias{sampler}
\title{main function for MCMC sampling designed for boolean matrix factorization}
\usage{
sampler(dat, model_options, mcmc_options)
}
\arguments{
\item{dat}{multivariate binary data (row for observations and column for dimensions4)}

\item{model_options}{Specifying assumed model options:
\itemize{
\item \code{n} The number of subjects.
\item \code{t_max} The maximum (guessed) number of clusters in the data during
posterior inference
\item \code{m_max} The maximum (guessed) number of factors during
the posterior inference
\item \code{a_theta, a_psi} hyperparameters for true and false positive rates;
a_theta and a_psi are both a vector of length two.
\item \code{log_v} The charaster string representing the prior
distribution for the number of true clusters, e.g.,
\code{"function(k) {log(0.1) + (k-1)*log(0.9)}"}. It is usually pre-computed log of the coefficients in Mixture of Finite Mixtures
(Miller and Harrison, 2017, JASA). Use this code:
\code{coefficients(eval(parse(text=model_options0$log_pk)),
model_options0$gamma,
model_options0$n,
model_options0$t_max+1)}
}
The following are used if one needs to prefix a few unnown parameters to
their respective true or other values
\itemize{
\item \code{Q} Q matrix
\item \code{theta} a vector of true positive rates
\item \code{psi} a vector of false positive rates
\item \code{p} a vector of machine prevalences
}
options for specifying data, sample size, max cluster number,
coefficients in MFM (Miller and Harrison 2017 JASA), gamma parameter in the MFM
Dirchlet prior, number of intermediate Gibbs scan to arrive at the launch state,
and other hyperparamter specification if needed, n_total for total number of
mcmc iterations and n_keep for the number of samples kept for posterior inference.
Note that the options involve other parameters for sampling hyperparameters such as
alpha in the Indian Buffet Process.}

\item{mcmc_options}{Options for MCMC sampling:
\itemize{
\item \code{n_total} total number of MCMC iterations
\item \code{n_keep} number of iterations kept
\item \code{n_split} the number of restricted Gibbs scan to arrive at a launch state;
see \link{restricted_gibbs}
\item \code{block_update_Q} update columns of Q or not. Then no constraints
about identifiability is imposed upon Q at any iterations.
}}
}
\value{
posterior samples for quantities of interest. It is a list comprised of the following elements
\itemize{
\item \code{t_samp}
\item \code{z_samp}
\item \code{N_samp}
\item \code{keepers} indices of MCMC samples kept for inference
\item \code{H_star_samp}
\item \code{alpha_samp}
} The following are recorded if they are not fixed in a priori:
\itemize{
\item \code{Q_samp}
\item \code{theta_samp}
\item \code{psi_samp}
\item \code{p_samp}
}
}
\description{
This function perform iterations over MCMC sampling chain.
NB: 1) add flexibility to specify other parameters as fixed. 2) sample component-specific
parameters. 3) sample other model parameters. 4) add timing and printing functionality.
5) add posterior summary functions.
}
