% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{scatterhist}
\alias{scatterhist}
\title{Plot posterior sampling chain and the histogram on the right margin}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{numeric vector; usually iteration ids}

\item{y}{the values at each x value}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}
}
\value{
a plot with sampling chain at the bottom left and histogram
on the right margin
}
\description{
NB: currently works for integer y; need to make it general.
}
\examples{
x <- 1:1000
y <- sample(1:10,1000,prob = c(1,2,3,4,5,5,4,3,2,1), replace=TRUE)
scatterhist(x,y)

}
