% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{simulate_Q}
\alias{simulate_Q}
\title{Randomly generate a Q matrix within the identifying conditions}
\usage{
simulate_Q(M, L, p = 0.1)
}
\arguments{
\item{M}{latent state dimension}

\item{L}{dimension of the binary responses}

\item{p}{Bernoulli probability of 1 in the Q matrix (except two diagonal matrices)}
}
\value{
a binary matrix of dimension M by L
}
\description{
This function initializes Q (if unknown) for MCMC sampling within identifiability
constraints
}
\examples{

simulate_Q(3,100)

}
