% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{simulate_Q_dat}
\alias{simulate_Q_dat}
\title{Randomly generate a Q matrix within the identifying conditions driven by data}
\usage{
simulate_Q_dat(M, dat, p = 0.1, frac = 1/5)
}
\arguments{
\item{M}{the number of factors (machines)}

\item{dat}{a matrix of binary data (rows for observations, columns for dimensions)}

\item{p}{the Bernoulli probability of an one in the Q matrix}

\item{frac}{only initialize those dimension with at least \code{frac}*100%
of subjects having a positive response.}
}
\value{
a binary matrix of dimension M by L
}
\description{
This function is used to initialize the MCMC sampling chain within identifiability
cosntraints; but this is smart - because it will not assign a one to a dimension with
few ones in the data
}
\examples{

#' # simulate data:
L0 <- 100
options_sim0  <- list(N = 200,  # sample size.
                     M = 3,   # true number of machines.
                     L = L0,   # number of antibody landmarks.
                     K = 8,    # number of true components.,
                     theta = rep(0.8,L0), # true positive rates
                     psi   = rep(0.01,L0), # false positive rates
                     alpha1 = 1 # half of the people have the first machine.
)

#image(simulate_data(options_sim0,SETSEED = TRUE)$datmat)
simu     <- simulate_data(options_sim0, SETSEED=TRUE)
simu_dat <- simu$datmat
simulate_Q_dat(5,simu_dat)

}
