% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoteTPhi.R
\name{smoteTPhi}
\alias{smoteTPhi}
\title{Temporal and Relevance Biased SMOTE}
\usage{
smoteTPhi(form, data, rel = "auto", thr.rel = 0.5, C.perc = "balance",
  k = 5, repl = FALSE, dist = "Euclidean", p = 2)
}
\arguments{
\item{form}{a model formula}

\item{data}{the original training set (with the unbalanced distribution)}

\item{rel}{is the relevance determined automatically (default: "auto") or provided by the user through a matrix. See examples.}

\item{thr.rel}{is the relevance threshold above which a case is considered as an extreme value}

\item{C.perc}{is a list containing the over-sampling percentage/s to apply to
all/each "class" obtained with the relevance threshold. The percentage represents the percentage of replicas that are added. Replicas of the examples are added randomly in each "class". Moreover, different percentages may be provided for each "class". Alternatively, it may be "balance" (the default) or "extreme", cases where the over-sampling percentages are automatically estimated.}

\item{k}{is the number of neighbours to consider as the pool from where the new generated examples are generated}

\item{repl}{is it allowed to perform sampling with replacement (bootstrapping).}

\item{dist}{is the distance measure to be used (defaults to "Euclidean"). Use "HEOM" if there are nominal and numerical predictors}

\item{p}{is a parameter used when a p-norm is computed}
}
\value{
a new training data set resulting from the application of the resampling strategy
}
\description{
Temporal and Relevance Biased SMOTE
}
\examples{
library(rewind)
data(temp)
ds <- create.data(temp,10)
C.perc <- list(4,0.5,4)
smoteTPhi <- smoteTPhi(V10 ~ ., ds, C.perc=C.perc)
smoteTPhi.Bal <- smoteTPhi(V10 ~ ., ds, C.perc="balance")
smoteTPhi.Ext <- smoteTPhi(V10 ~ ., ds, C.perc="extreme")

}
