% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{summarize_latent_state_given_Q}
\alias{summarize_latent_state_given_Q}
\title{Summarize individual latent state patterns given Q}
\usage{
summarize_latent_state_given_Q(H_star_samp, z_samp, Q)
}
\arguments{
\item{H_star_samp}{A latent state profile array with elements indexed by
(pseudo-cluster, latent state, MCMC iteration for inference)}

\item{z_samp}{A pseudo-cluster indicator matrix with elements indexed by (individual, MCMC
iteration for inference)}

\item{Q}{Q matrix; known but could have rows to be ordered}
}
\value{
\itemize{
\item \code{marg_prob} the
marginal probabilities of each individual's latent states being active
\item \code{H_pat_ordered_samp} the posterior samples of individual
 latent states (after ordering
by row of Q and converting the binary codes to decimal numbers)
}
}
\description{
This function takes in posterior samples of individual latent states,
orders latent states according to the rows of Q and output 
\itemize{
\item 1) the
marginal probabilities of each individual's latent states being active
and
\item 2) the posterior samples of individual latent states (after ordering
by row of Q and converting the binary codes to decimal numbers)
}
}
