% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_rbeta}
\alias{trunc_rbeta}
\title{Sample truncated Beta random variables}
\usage{
trunc_rbeta(n, a, b, lower = 0, upper = 1)
}
\arguments{
\item{n}{sample size}

\item{a}{first Beta parameter}

\item{b}{second Beta parameter}

\item{lower}{lower limit}

\item{upper}{upper limit}
}
\description{
Sample truncated Beta random variables
}
\examples{
hist(trunc_rbeta(100000,1,1,0.2,0.8),freq=FALSE,xlim=c(0,1));abline(h=1.67)
hist(trunc_rbeta(100000,2,3,0.2,0.8),freq=FALSE,xlim=c(0,1))
}
