% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{update_mu_PR}
\alias{update_mu_PR}
\title{sample a_theta_one - hyperparameter for positive rates theta_l, psi_l (mean)}
\usage{
update_mu_PR(theta_vec, s, a = 2, b = 1, show_density = TRUE)
}
\arguments{
\item{theta_vec}{vector of positive rates}

\item{s}{a positive scalar value; represent the current sampled value of pseudo-sample size}

\item{a, b}{hyperparameter for beta distribution over original a_theta_one.}

\item{show_density}{Default to FALSE - hide the full conditional density of alpha_s
given other unknown parameters; TRUE otherwise.}
}
\value{
an updated alpha_s value (positive)
}
\description{
Function to sample the parameters from a grid
}
\seealso{
\code{\link{update_pseudo_n_PR}} for examples.
}
