% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{update_positive_rate}
\alias{update_positive_rate}
\title{update true and false positive rates}
\usage{
update_positive_rate(Y, H, Q, a_theta, a_psi, old_theta_vec = NULL)
}
\arguments{
\item{Y}{binary data matrix}

\item{H}{a binary matrix: rows for all individuals, columns for latent states}

\item{Q}{Q matrix}

\item{a_theta, a_psi}{Beta hyperparameters for priors on \code{theta} and
\code{psi}, respectively. 2 by L matrices.}

\item{old_theta_vec}{sampled \code{theta} parameters from the previous iteration (vector
of length L); to sample \code{psi} given the constraint}
}
\value{
a list of true \code{theta} and false \code{psi} positive rates, each of length L.
}
\description{
This function samples the true (theta) and false (psi) positive rates;
Constraint: \code{theta[l]> 0.5 > psi[l]}
}
