% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{update_prevalence}
\alias{update_prevalence}
\title{update prevalence}
\usage{
update_prevalence(H_star, alpha, M)
}
\arguments{
\item{H_star}{machine usage profiles for all clusters}

\item{alpha}{hyperparameter}

\item{M}{total number of machines (number of columns for H_star)}
}
\value{
vector of prevalences of length M.
}
\description{
This function updates the prevelance paramters
}
