% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{update_Q}
\alias{update_Q}
\title{update the Q matrix element-wise}
\usage{
update_Q(Y, Q_old, H, z, t, mylist, p, theta, psi, constrained = FALSE)
}
\arguments{
\item{Y}{binary data matrix}

\item{Q_old}{the Q matrix from the last scan in Gibbs sampler (of dimension M by L)}

\item{H}{matrix of latent state profiles, rows for N subjects,
columns for M latent states}

\item{z}{a vector of individual cluster indicators}

\item{t}{the number of clusters at a particular iteration}

\item{mylist}{the ordered list that keeps the cluster ids in increasing order}

\item{p}{Latent state prevalences; a vector of length identical to the columns of H.}

\item{theta, psi}{True and false positive rates. Both are vectors of length L}

\item{constrained}{Default to FALSE; Set to TRUE if doing constrained Gibbs sampling
with identifiability constraints.}
}
\value{
A Q matrix with all its elements updated.
}
\description{
This function updates Q matrix by Gibbs sampler (with option to do constarined
updates within the identifiability constraint)
NB: - do we need M here? do we modify M after collapsing partner machines.
    - need to speed up.
}
