% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewind_junk.R
\name{update_Q_no_H}
\alias{update_Q_no_H}
\title{update the Q matrix element by element}
\usage{
update_Q_no_H(Y, Q_old, z, t, mylist, p, theta, psi)
}
\arguments{
\item{Y}{data}

\item{Q_old}{the Q matrix from the last scan in Gibbs sampler (of dimension M by L)}

\item{p}{prevalence of machines; a vector of length identical to the columns of H.}

\item{theta, psi}{True and false positive rates. Both are vectors of length L}

\item{H}{matrix of machine usage indicators, rows for N subjects, columns for M machines}
}
\value{
\itemize{
\item \code{datmat} a matrix for multivariate binary observations from the assumed boolean
matrix factorization,
\item \code{Q} A Q-matrix,
\item \code{H_star} A binary matrix of dimension K by M, for K clusters, and M factors,
\item \code{Z} A vector of length N for individual cluster indicators,
\item \code{xi} a binary matrix of dimension N by L for true presence or absence of proteins
\item \code{Eta} A binary matrix of dimension N by M for presence or absence of factors among
all subejcts.
}
}
\description{
this function updates Q matrix by constrained Gibbs sampler
NB: - do we need M here? do we modify M after collapsing partner machines.
    - need to speed up.
}
