% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{z2comat}
\alias{z2comat}
\title{Convert the cluster indicators to the co-activation patterns.}
\usage{
z2comat(z)
}
\arguments{
\item{z}{a vector of integers indicating cluster assignments for all subjects}
}
\value{
an N by N binary; an 1 represents that two subjects are in the same
cluster; 0 otherwise.
}
\description{
Convert the cluster indicators to the co-activation patterns.
}
\examples{
z2comat(c(1,1,2,2,3,4,5,6,5,7))
}
