% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{do_update_Q_one}
\alias{do_update_Q_one}
\title{check whether a vector is equal to a unit vector with the one at a particular
position}
\usage{
do_update_Q_one(Q, k, l)
}
\arguments{
\item{Q}{a matrix with rows being machines and columns being protein landmarks (dimension)}

\item{k}{the index that is being checked if \code{v[k]} is the only one in
vector \code{v}. \code{k} must be smaller than or equal to the length of k}

\item{v}{the vector (a binary vector)}

\item{Q}{a matrix with rows being machines and columns being protein landmarks (dimension)}

\item{k, l}{row and column indicators for checking wether to update Q_kl}
}
\value{
true for \eqn{v = \mathbf{e}_k}

a matrix filled with logical values of dimensions identical to Q. TRUE for updating
in constrained Gibbs sampler, FALSE for skipping the updating.
determine to update the current element of Q_ml or not

Function to test whether we need to update the current element of Q_ml. This
is needed in the constrained Gibbs sampler.

logical value. TRUE for updating in constrained Gibbs sampler, FALSE
for skipping the updating.
}
\description{
check whether a vector is equal to a unit vector with the one at a particular
position
}
\examples{
determine to update the current element of Q_ml or not


Function to test whether we need to update the current element of Q_ml. This
is needed in the constrained Gibbs sampler.

}
