% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{log_pr_Qml_cond}
\alias{log_pr_Qml_cond}
\title{compute the full conditional probability of Q_ml given others}
\usage{
log_pr_Qml_cond(Yl, eta_star, Ql, thetal, psil)
}
\arguments{
\item{Yl}{a column of the multivariate binary data}

\item{eta_star}{a matrix of machine usage indicators, rows for clusters, columns for M machines}

\item{Ql}{l-th column of the Q matrix}

\item{thetal, psil}{True and false positive rate (both are numbers between 0 and 1)}

\item{p}{prevalence of machines; a vector of length identical to the columns of H.}
}
\value{
log conditional probability of Q_ml given other unknown parameters
}
\description{
Function to compute this log full conditional density, which will be used
in \link{update_Q}
}
