% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{log_pr_Qml_cond_enum}
\alias{log_pr_Qml_cond_enum}
\title{Compute the full conditional probability of column Q_l given other unknown parameters}
\usage{
log_pr_Qml_cond_enum(Yl, eta_star, Ql_enumerate, thetal, psil)
}
\arguments{
\item{Yl}{a column of the multivariate binary data}

\item{eta_star}{a matrix of machine usage indicators, rows for clusters, columns for M machines}

\item{Ql_enumerate}{l-th column of the Q matrix}

\item{thetal, psil}{True and false positive rates. Both are vectors of length L}
}
\value{
a vector of log conditional probability of column Q_l taking each of
\code{2^nrow(Ql_enumerate)} given other unknown parameters,
the dimension will be identical to \code{2^nrow(Ql_enumerate)}.
}
\description{
Function to compute the log full conditional density for all patterns of the
l-th column of Q. This function is used in \link{update_Q_col_block}.
}
