% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{merge_H_Q}
\alias{merge_H_Q}
\title{Merge latent state profile matrix (H_star) and Q matrix}
\usage{
merge_H_Q(H_star_redun, mylist, t, Q, VERBOSE = FALSE, z_pseudo = NULL,
  skip_Q = FALSE)
}
\arguments{
\item{H_star_redun}{A binary matrix of rows \code{t_max+3} by \code{m_max};
It may be redundant because it has rows corresponding to empty pseudo-clusters,
and if the entire matrix is non-zero, the zero columns mean inactive latent states}

\item{mylist}{Cluster labels; A vector of length \code{t_max+3} comprised of
integers that need not be consecutive.}

\item{t}{The number of non-empty pseudo-clusters}

\item{Q}{The \code{m_max} by \code{L} Q matrix}

\item{VERBOSE}{Default to \code{FALSE}: no print of the merged matrices; Otherwise
set to \code{TRUE}.}

\item{z_pseudo}{Default is \code{NULL}.
A vector of pseudo-cluster indicators to be merged into
scientific cluster indicators \code{z_sci}.}

\item{skip_Q}{Default is \code{FALSE} - merge Q rows with partner states;
Set to \code{TRUE} if \code{Q} is given.}
}
\value{
A list comprised of two elements named:
\code{H_star_merge} and \code{Q_merge} and \code{z_sci} if \code{z_pseudo}
is provided.
}
\description{
This function
\itemize{
\item merges identical rows in H: pseudo-cluster ---> scientific clusters;
\item merges partner latent states (identical non-zero columns in H);
merge the corresponding rows in Q by \code{pmax}.
}
}
\examples{

# the third latent state is inactive among non-empty pseudo-clusters;
# the 2nd and 4th latent states are partners.
# merge 1 and 2 pseudo-clusters.
 H_star_redun <- matrix(c(0,1,0,1,
                          0,1,0,1,
                          1,0,0,0,
                          0,0,0,0,
                          0,0,0,0,
                          0,0,0,0,
                          0,0,0,0),nrow=7,ncol=4,byrow=TRUE)
mylist <- c(1,2,3,5,0,0,0)
t <- 4
Q <- simulate_Q(4,100,p=0.1)
z_pseudo <- c(1,1,3,5,2,2,2,1,3,3,5,1,2,2,1)
merge_H_Q(H_star_redun,mylist,t,Q,TRUE,z_pseudo)

# if all zero columns in H_star exists, they will be ignored.

}
