% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{merge_map}
\alias{merge_map}
\title{Merge mapping}
\usage{
merge_map(pseudo_pat, uniq_pat)
}
\arguments{
\item{pseudo_pat}{a vector of character strings for
binary patterns that might not be unique}

\item{uniq_pat}{a vector of character strings of distinct binary codes}
}
\value{
a list \itemize{
\item \code{map} a vector of integer of identical length to \code{pseudo_pat}; takes values
from 1 to \code{length(uniq_pat)}
\item \code{uniq_pat} a matrix of unique binary patterns (# rows =  \code{length(uniq_pat)},
# columns = number of 1/0s for each element in \code{uniq_pat})
}
}
\description{
This function combines pseudo-clusters by the current H_star samples.
}
\examples{
#' # simulate data:
L0 <- 100
options_sim0  <- list(N = 200,  # sample size.
                     M = 3,   # true number of machines.
                     L = L0,   # number of antibody landmarks.
                     K = 8,    # number of true components.,
                     theta = rep(0.8,L0), # true positive rates
                     psi   = rep(0.01,L0), # false positive rates
                     alpha1 = 1 # half of the people have the first machine.
)

#image(simulate_data(options_sim0,SETSEED = TRUE)$datmat)

simu     <- simulate_data(options_sim0, SETSEED=TRUE)
tmp <-  simu$H_star[c(1,1,2,2,2,3,4,5,6,7,8),]
uid <- unique(apply(tmp,1,paste,collapse=""))
merge_map(apply(tmp,1,paste,collapse=""),uid)
}
