% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{metrop_flip}
\alias{metrop_flip}
\title{Metroplis flipping to improve mixing}
\usage{
metrop_flip(x, curr_prob)
}
\arguments{
\item{x}{current value, 0 or 1}

\item{curr_prob}{current probability of being 1}
}
\value{
a binary value
}
\description{
Metroplis flipping to improve mixing
}
\examples{

metrop_flip(1,0.2)
}
