% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{plot_individual_pred}
\alias{plot_individual_pred}
\title{Plot individual probabilities}
\usage{
plot_individual_pred(p_samp, state_nm, thedata)
}
\arguments{
\item{p_samp}{A vector of posterior probabilities for each individual}

\item{state_nm}{A string vector for the names of the latent state dimensions}

\item{thedata}{a vector of binary measurements for a subject; of identical
length to `state_nm`}
}
\value{
a figure with population etiology fraction information (
NB: need to add more info.)
}
\description{
Plot individual probabilities
}
