% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate data from boolean matrix factorization model}
\usage{
simulate_data(options_sim, SETSEED = FALSE)
}
\arguments{
\item{options_sim}{options for simulating data: N - sample size; M - number of machines;
L - number of protein landmarks; K - number of ture components; alpha1 - hyperparameter
for determining the machine prevalence {p_m}; theta - true positive rates for
all protein landmarks; psi - false positive rates for all protein landmarks.}

\item{SETSEED}{FALSE by default; Set to TRUE if to fix random number generation.}
}
\value{
\itemize{
\item \code{datmat} a matrix for multivariate binary observations from the assumed boolean
matrix factorization,
\item \code{Q} A Q-matrix,
\item \code{H_star} A binary matrix of dimension K by M, for K clusters, and M factors,
\item \code{Z} A vector of length N for individual cluster indicators,
\item \code{xi} a binary matrix of dimension N by L for true presence or absence of proteins
\item \code{Eta} A binary matrix of dimension N by M for presence or absence of factors among
all subejcts.
}
}
\description{
This function takes in options_sim and simulate as specified by the options.
}
\examples{
# simulate data:
L0 <- 100
options_sim0  <- list(N = 200,  # sample size.
                     M = 3,   # true number of machines.
                     L = L0,   # number of antibody landmarks.
                     K = 8,    # number of true components.,
                     theta = rep(0.8,L0), # true positive rates
                     psi   = rep(0.01,L0), # false positive rates
                     alpha1 = 1 # half of the people have the first machine.
)

image(simulate_data(options_sim0,SETSEED = TRUE)$datmat)
simu     <- simulate_data(options_sim0, SETSEED=TRUE)
simu_dat <- simu$datmat
}
