% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{simulate_Q}
\alias{simulate_Q}
\title{Randomly generate a Q matrix within the identifying conditions}
\usage{
simulate_Q(M, L, p = 0.1)
}
\arguments{
\item{M}{the number of factors (machines)}

\item{L}{the dimension of the binary data}

\item{p}{the Bernoulli probability of an one in the Q matrix}
}
\value{
a binary matrix of dimension M by L
}
\description{
This function is used to initialize the MCMC sampling chain within identifiability
cosntraints
}
\examples{

simulate_Q(3,100)

}
