% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{simulate_Q_dat}
\alias{simulate_Q_dat}
\title{Randomly generate a Q matrix within the identifying conditions driven by data}
\usage{
simulate_Q_dat(M, dat, p = 0.1)
}
\arguments{
\item{M}{the number of factors (machines)}

\item{dat}{a matrix of binary data (rows for observations, columns for dimensions)}

\item{p}{the Bernoulli probability of an one in the Q matrix}
}
\value{
a binary matrix of dimension M by L
}
\description{
This function is used to initialize the MCMC sampling chain within identifiability
cosntraints; but this is smart - because it will not assign a one to a dimension with
few ones in the data
}
\examples{

simu     <- simulate_data(options_sim0, SETSEED=TRUE)
simu_dat <- simu$datmat
simulate_Q_dat(5,simu_dat)

}
