% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{update_alpha}
\alias{update_alpha}
\title{sample alpha - hyperparameter for prevalences p_m}
\usage{
update_alpha(H_star, t, M, a = 1, b = 1, show_density = FALSE)
}
\description{
Function to sample the parameters from a grid
}
\details{
@param H_star the matrix of machine usage profiles across clusters
 @param t number of clusters
 @param M number of machines
 @param a,b hyperparameter for beta distribution over reparameterized alpha.
 @param show_density Default to FALSE - not showing the full conditional density of alpha
 given other unknown parameters; TRUE otherwise.
}
