% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{update_positive_rate}
\alias{update_positive_rate}
\title{update true and false positive rates WITHOUT constraints}
\usage{
update_positive_rate(Y, H, Q, a_theta, a_psi)
}
\arguments{
\item{Y}{data}

\item{H}{a matrix whose rows are individual specific machine indicators (all subjects)}

\item{Q}{Q matrix}

\item{a_theta, a_psi}{hyperparameters for priors on theta and psi, respectively}
}
\value{
a list of true positive rates theta and false positive rates psi, each of length L.
}
\description{
This function samples the theta and psi parameters in bmf
}
