% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mcmc.R
\name{update_prevalence}
\alias{update_prevalence}
\title{update latent state prevalences}
\usage{
update_prevalence(H_star, alpha, M)
}
\arguments{
\item{H_star}{latent state profiles for all clusters}

\item{alpha}{hyperparameter}

\item{M}{dimension of the latent state vector (number of columns for H_star)}
}
\value{
vector of length M (each between 0 and 1).
}
\description{
This function updates the prevalence parameters
}
