% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{update_Q_col_block}
\alias{update_Q_col_block}
\title{Block update the Q matrix column by column}
\usage{
update_Q_col_block(Y, Q_old, H, z, t, mylist, p, theta, psi)
}
\arguments{
\item{Y}{data}

\item{Q_old}{the Q matrix from the last scan in Gibbs sampler (of dimension M by L)}

\item{H}{matrix of machine usage indicators, rows for N subjects, columns for M machines}

\item{z}{a vector of individual cluster indicators}

\item{t}{the number of clusters at an iteration}

\item{mylist}{the ordered list that keeps the cluster ids in increasing order}

\item{p}{prevalence of machines; a vector of length identical to the columns of H.}

\item{theta, psi}{True and false positive rates. Both are vectors of length L}
}
\value{
A Q matrix with all elements updated.
}
\description{
This function updates Q matrix by Gibbs sampler (with option to do constarined
updates within identifiability constraint or not)
NB: - do we need M here? do we modify M after collapsing partner machines.
    - need to speed up.
}
