% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{z2comat}
\alias{z2comat}
\title{Compute the posterior co-clustering probability matrix (probability that i and j
are clustered together).}
\usage{
z2comat(z)
}
\arguments{
\item{z}{a matrix of posterior samples, with subjects and MCMC samples in
the rows and columns, respectively.}
}
\value{
a matrix of empirical co-clustering frequencies based on the
posterior samples
}
\description{
This function is to evaluate the recovered clusters
}
\examples{
z2comat(matrix(c(1,1,2,2,3,4,5,6,5,7),ncol=1))

}
